% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello()
}
\description{
You do not need to unit test this function
}
\details{
This function should exist in the NAMESPACE
}
\examples{

# this calls the function and should be working
hello_dsci310_class()
}
