% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_activ.R
\name{recode_activ}
\alias{recode_activ}
\title{recode_activ}
\usage{
recode_activ(activ_value)
}
\arguments{
\item{activ_value}{The active value given}
}
\value{
the active state currently in
}
\description{
Determine the active value of the given variable
}
\examples{
activ1 <- recode_activ(1);
activ2 <- recode_activ();
activ3 <- recode_activ("hello");
}
